Refactor NeoFaker.EnUs.Person module
